Refactor: changed ids to longs in model

+++ app/src/main/java/com/beesham/shopifymerchantstore/model/Image.java
@@ -9,10 +9,10 @@ public class Image {
 
     @SerializedName("id")
     @Expose
-    private int id;
+    private long id;
     @SerializedName("product_id")
     @Expose
-    private int productId;
+    private long productId;
     @SerializedName("position")
     @Expose
     private int position;
@@ -35,19 +35,19 @@ public class Image {
     @Expose
     private List<Object> variantIds = null;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
